just moved the history.cs and Robot.cs into a Models folder
