+++ BlazorCRUD/Components/Pages/HistoryPages/Create.razor
@@ -1,6 +1,7 @@
 ﻿@page "/history/create"
 @inject BlazorCRUD.Data.ApplicationDbContext DB
-@using BlazorCRUD
+@    using BlazorCRUD;
+    using BlazorCRUD.Models
 @inject NavigationManager NavigationManager
 
 <PageTitle>Create</PageTitle>
