+++ BlazorCRUD/Components/Pages/HistoryPages/Delete.razor
@@ -1,9 +1,9 @@
 ﻿@page "/history/delete"
 @inject BlazorCRUD.Data.ApplicationDbContext DB
-@using BlazorCRUD
+@    using BlazorCRUD
 @inject NavigationManager NavigationManager
-@using Microsoft.EntityFrameworkCore
-@using Microsoft.AspNetCore.Authorization
+@    using Microsoft.EntityFrameworkCore
+@    using Microsoft.AspNetCore.Authorization
 
 <PageTitle>Delete</PageTitle>
