+++ BlazorCRUD/Components/Pages/HistoryPages/Edit.razor
@@ -1,8 +1,8 @@
 ﻿@page "/history/edit"
 @inject BlazorCRUD.Data.ApplicationDbContext DB
-@using BlazorCRUD
+@    using BlazorCRUD
 @inject NavigationManager NavigationManager
-@using Microsoft.EntityFrameworkCore
+@    using Microsoft.EntityFrameworkCore
 
 <PageTitle>Edit</PageTitle>
