+++ BlazorCRUD/Models/Robot.cs
